refactor: use scss variables and restructure components folders
