+++ pricing-react/src/components/Cart/FilledCart/CartItem/CartItem.module.scss
@@ -1,4 +1,4 @@
-@import "../../../variables.scss";
+@import "../../../../variables.scss";
 
 .cart-item {
   background-color: $primary-050;
