+++ pricing-react/src/components/Cart/FilledCart/CartSummary/CartSummary.module.scss
@@ -1,4 +1,4 @@
-@import "../../../variables.scss";
+@import "../../../../variables.scss";
 
 .summary {
   background-color: $primary-050;
