+++ pricing-react/src/components/Cart/FilledCart/ItemSummary/ItemSummary.module.scss
@@ -1,4 +1,4 @@
-@import "../../../variables.scss";
+@import "../../../../variables.scss";
 
 .container {
   display: flex;
