+++ pricing-react/src/components/Icons/PricingIcon/PricingIcon.module.scss
@@ -1,5 +1,4 @@
-$complementary-600: #ea580c;
-$primary-050: #fcfcfc;
+@import "../../../variables.scss";
 
 .pricing-icon {
   rect {
